src/requirements.txt
# aiofiles==23.2.1
# aiohappyeyeballs==2.4.4
# aiohttp==3.11.11
# aiosignal==1.3.2
# annotated-types==0.7.0
# anthropic==0.43.0
# anyio==4.8.0
# asyncer==0.0.7
# asyncpg==0.30.0
# attrs==24.3.0
# azure-core
# azure-storage-blob
# backoff==2.2.1
# bidict==0.23.1
# certifi==2024.12.14
# cffi==1.17.1
# chainlit==2.0.3
# charset-normalizer==3.4.1
# chevron==0.14.0
# click==8.1.8
# colorama==0.4.6
# cryptography==44.0.0
# dataclasses-json==0.6.7
# Deprecated==1.2.15
# distro==1.9.0
# fastapi==0.115.6
# filetype==1.2.0
# frozenlist==1.5.0
# googleapis-common-protos==1.66.0
# grpcio==1.69.0
# h11==0.14.0
# httpcore==1.0.7
# httpx==0.28.1
# idna==3.10
# importlib_metadata==8.5.0
# inflection==0.5.1
# isodate==0.7.2
# Jinja2==3.1.5
# jiter==0.8.2
# Lazify==0.4.0
# literalai==0.1.103
# MarkupSafe==3.0.2
# marshmallow==3.25.1
# monotonic==1.6
# multidict==6.1.0
# mypy-extensions==1.0.0
# nest-asyncio==1.6.0
# openai==1.59.7
# opentelemetry-api==1.29.0
# opentelemetry-exporter-otlp==1.29.0
# opentelemetry-exporter-otlp-proto-common==1.29.0
# opentelemetry-exporter-otlp-proto-grpc==1.29.0
# opentelemetry-exporter-otlp-proto-http==1.29.0
# opentelemetry-instrumentation==0.50b0
# opentelemetry-instrumentation-alephalpha==0.36.0
# opentelemetry-instrumentation-anthropic==0.36.0
# opentelemetry-instrumentation-bedrock==0.36.0
# opentelemetry-instrumentation-chromadb==0.36.0
# opentelemetry-instrumentation-cohere==0.36.0
# opentelemetry-instrumentation-google-generativeai==0.36.0
# opentelemetry-instrumentation-groq==0.36.0
# opentelemetry-instrumentation-haystack==0.36.0
# opentelemetry-instrumentation-lancedb==0.36.0
# opentelemetry-instrumentation-langchain==0.36.0
# opentelemetry-instrumentation-llamaindex==0.36.0
# opentelemetry-instrumentation-logging==0.50b0
# opentelemetry-instrumentation-marqo==0.36.0
# opentelemetry-instrumentation-milvus==0.36.0
# opentelemetry-instrumentation-mistralai==0.36.0
# opentelemetry-instrumentation-ollama==0.36.0
# opentelemetry-instrumentation-openai==0.36.0
# opentelemetry-instrumentation-pinecone==0.36.0
# opentelemetry-instrumentation-qdrant==0.36.0
# opentelemetry-instrumentation-replicate==0.36.0
# opentelemetry-instrumentation-requests==0.50b0
# opentelemetry-instrumentation-sagemaker==0.36.0
# opentelemetry-instrumentation-sqlalchemy==0.50b0
# opentelemetry-instrumentation-threading==0.50b0
# opentelemetry-instrumentation-together==0.36.0
# opentelemetry-instrumentation-transformers==0.36.0
# opentelemetry-instrumentation-urllib3==0.50b0
# opentelemetry-instrumentation-vertexai==0.36.0
# opentelemetry-instrumentation-watsonx==0.36.0
# opentelemetry-instrumentation-weaviate==0.36.0
# opentelemetry-proto==1.29.0
# opentelemetry-sdk==1.29.0
# opentelemetry-semantic-conventions==0.50b0
# opentelemetry-semantic-conventions-ai==0.4.2
# opentelemetry-util-http==0.50b0
# packaging==23.2
# plotly==5.24.1
# posthog==3.8.3
# propcache==0.2.1
# protobuf==5.29.3
# pycparser==2.22
# pydantic==2.10.5
# pydantic_core==2.27.2
# PyJWT==2.10.1
# python-dateutil==2.9.0.post0
# python-dotenv==1.0.1
# python-engineio==4.11.2
# python-multipart==0.0.18
# python-socketio==5.12.1
# regex==2024.11.6
# requests==2.32.3
# simple-websocket==1.1.0
# six==1.17.0
# sniffio==1.3.1
# starlette==0.41.3
# syncer==2.0.3
# tenacity==8.5.0
# tiktoken==0.8.0
# tomli==2.2.1
# tqdm==4.67.1
# traceloop-sdk==0.36.0
# typing-inspect==0.9.0
# typing_extensions==4.12.2
# uptrace==1.27.0
# urllib3==2.3.0
# uvicorn==0.25.0
# watchfiles==0.20.0
# wrapt==1.17.2
# wsproto==1.2.0
# yarl==1.18.3
# zipp==3.21.0
#david
chainlit
aiohttp
python-dotenv
pytest
openai
azure-ai-projects
azure-ai-agents
# azure-ai-foundry
azure-identity
mcp
semantic-kernel
pydantic